+++ Tasks/OrchestratorV2/interfaces/task/endpoint.ts
@@ -1,7 +1,6 @@
 export interface IEndpoint {
 
     url: string;
-    account: string;
     token: string;
 
 }